Make detach delete atomic
The problem was that between reading and deleting the relationships
associated to the node to be detached deleted another transaction
could have added new relationships which would have missed during the
read and hence not removed.  Such race would cause the transaction
performing the detach delete operation to fail when trying to delete
the node.  This change will make sure that detach delete is somewhat
atomic.  Now we collect all the involved nodes and we lock them (by id
order) and after that we try to reread the relationship iterator, if
it is the same than we'll continue by deleting all the relationships
and the node, otherwise we'll release all the acquired locks and try
again.