Packer should not fail on deleted entities
When returning deleted entitities such as in `CREATE (a) DELETE a RETURN a`
we should still be able to pack and return the deleted entity to the user. Note
that in the other APIs we currently also mark the entity as deleted, this PR does not
add that to BOLT but that will come at a later stage.