Increase default backup timeout
This timeout is basically Netty's blocking read timeout. It limits amount of
time backup waits to read the next chunk from a channel. After the timeout
BlockingReadHandler throws BlockingReadTimeoutException.

Backup triggers a checkpoint which triggers a store flush. For large stores
this flush could take quite a bit of time especially if IO throttling is turned
on in the checkpointer.

This commit makes default backup timeout 20min instead of 40s.