Revert change in NiceHasher causing regression
Profiling showed that how the sets of keys and values are combined for
Map values in NiceHasher has a performance impact on some queries,
e.g. Cypher benchmark Levelstory Q10. The regression was introduced by
a commit to fix warnings 799925074feddfde1db8e1737ad19679223aec97.
This only changes the additional code with an impact and does not
affect the warning.