Remove the command prompt functionality from Neo4j Desktop
This is broken by changes elsewhere. It is as yet undecided whether the
problems will be fixed and the functionality reinstated or whether it is
removed permanently in favour of using the tarball or zip for advanced
functionality like Neo4j Shell; therefore this is a minimal change,
leaving unused code which will be cleaned up later.