Skip unused property records when reading chain
To be able to increase property read throughput our reads should be more
tolerant to concurrent updates of the same property chains. One of the
steps - skip already removed properties.

To do that store property cursor will skip unused records and will read and
build chain of only used records. Exception that record is not in used will not
be thrown anymore.

Co-authored-by: @MishaDemianenko