Rename the old dbms.querylog.filename setting to mark it as internal
This setting has been removed from the product surface. It's now an
internal, derived setting and should have been renamed to reflect that.

This fixes a bug where someone who used the query log and had customized
the name of the log file, but had not correctly migrated their
configuration, would end up with a server which would not start.