Introduce runtime conversion mechanism
At runtime we need to be able to convert to public types while doing the scala to java
conversion. Since we should not depend on external types directly within the compiler module
the actual conversion function must be defined outside the compiler module but applied within
the module. Currently this mechanism is only used for `Point` but could be extended to map things
as `Node`, `Relationship`, and `Path` as well.