Capture error messages during service install and uninstall with Powershell
This commit captures STDERR for PRUNSRV during Windows service install and
uninstall.  If there as an error, STDERR is written to the console.

Previously during service installation and uninstallation, a success message
is written to console even if it failed.  This commit outputs an appropriate
message depending on the exit code.