core-edge: Batching inbound message queue for RAFT.
The RAFT message processing is now running on a dedicated thread distinct from the
Netty networking thread. They are connected by a queue and new entry requests are
batched together before being processed by the leader. This leads to batching when
appending to the RAFT log and when sending new entries.