core-edge: batch append to segmented log
The interface of RaftLog is updated to handle an array of entries and
the BatchAppendLogEntries command is updated to utilize it.

Only the segmented log is actually updated to take advantage of it. The
other log implementations can easily be adapted later if required.

The contract test has been updated to utilize a single append call
with several entries where applicable.

The in-queue and max-batch of the RAFT were also both marked internal
and fixed at 64. They are still configurable but it seems likely
unnecessary in most all cases to do it, so no need to clutter the
interface with it.