Pass through NEO4J_HOME to tools called from neo4j-admin
This is needed so that reading of relative values from the
config file is done relative to Neo4j rather than relative
to where you call the script from.

Moved ConfigLoader across from the server package as that
can be fed the home-dir and will then work out relative
paths for us.