Fixing flaky CoreToCoreCopySnapshotIT.shouldBeAbleToDownloadToNewInstanceAfterPruning
If the last entry in a snapshot was not a transaction then the
CoreState would try to apply an entry from before the start of
the log.

Now we tell CoreState#skip that is has already applied up to the index
of the snapshot, rather than relying on the index from the transaction
log.