Slave - master lock session termination
Commit makes it possible to terminate lock sessions that slaves have on master.
This is useful when terminating write transactions on slaves because they have
corresponding lock clients on master.

When SlaveLocksClient is stopped it will end the corresponding lock session on
master and stop local locks client.

Feature is guarded by `experimental.tx_termination_aware_locks` settings which
is off by default.