Increase some page cache test timeouts
Most platforms implement `fsync` in a way that is a lot slower than on OS X (because the OS X version does nothing), so some of these test were running dangerously close to the test time out, and occasionally exceding it.
So this should make these tests not flaky anymore.