Reorder jvm args to match Java9 requirements
On Java9, the hashcode setting must come after the experimental
switch (this has no effect on other java versions).

This allows Neo4j to start and crash at the real place where we are not
compatible with Java9, instead of failing to start due to the order of
jvm arguments.