Make `ExternalResourceNotFound` a `ClientError`
The only usage of this status code is for when `LOAD CSV` is unable to
load its csv file parameter. Telling this to the client directly is more helpful
than referring them to the server log. The assumption here is that the client
has supplied an incorrect parameter (wrong file name).