Add a feature toggle to disable the schema lock
The schema lock is a little bit special in that the readers vastly outnumber the writers - even more so than usual - since schema changes are quite rare.
This feature toggle allows us to see the upper bound for the performance that we can gain by building a specialised implementation of the schema lock.