Optimise `PagedFile.flushAndForce`
Taking a flush lock on a page is fairly cheap, but it is not free; it involves a CAS loop to take the lock, and another to undo the lock.
If a file is very large, it is quite likely that the vast majority of its pages are not dirty.
Yet, we were taking flush locks on all the pages in turn, just to check their dirty bit.
This meant that a significant amount of time was spent locking pages, which can for very large files cause a rather dramatic slowdown in flushing performance.
We now check the dirty bit under an optimistic read lock, which still gives us the visibility we need when racing with eviction.
If the page is found to be dirty, we then take the flush lock and double check.
If the page is not dirty, then our read of the dirty bit did not race with eviction nor with any writes.
Flushing is allowed to race with writes anyway, and pages that are evicted are kept eviction-locked for the duration of their time as free pages.
This means that we can safely skip pages that are observed to be clean under our optimistic read lock.
The result of this optimisation is that we can now process large swaths of clean pages about twice as fast as before.