Make transaction id accessible after KTI commit
Transaction id is only assigned during commit of KernelTransactionImplementation.
It is not assigned if transaction was rolled back or read-only. Currently it is
not possible to obtain committed transaction id because KTI just implements
AutoCloseable and commit happens in #close().

However, it might be sometimes valuable to know committed transaction id and
track it.

This commit makes KTI expose it's id (if committed) via #closeTransaction()
method while still keeping KTI an AutoCloseable.