core-edge: core state now exposes last applied
Last applied index now also always counts content which is not
commands for the state machines, and thus an assertion which
relates to that had to be removed.

This is added so that an external observer can see if everything
in queue for being applied has been applied, by comparing the
index to the appendIndex of the log.