CsvInput parses data in parallel
takes advantage of being able to determine where any data row starts
and another ends if multi-line fields are disallowed. In this case
chunks of csv data can be read and handed off to one or more threads
parsing the data, to finally end up merged back in the same order as
they were read from the csv source.
From the outside there's no difference, it still provides `InputIterator`
for nodes and relationships.