Don't take locks for entities created in the same transaction.
There is no need to take locks for new entities that have not yet been
committed because they cannot yet be visible to any other transaction.

There should have been no contention for these locks, so this change
should not have a significant impact, except where the cost of acquiring
any lock is high, for example when acquiring locks for a slave in an HA
cluster, where each lock acquisition requires a network call.