Fix buffer leak in Bolt output
BoltProtocolV1 is a container for inbound and outbound Netty-based connections.
There exists a protocol instance for each neo4j client. Previously outbound
connection, represented by ChunkedOutput class, was passed in as a parameter
and never closed. This caused a native memory leak because connections
allocate native byte buffers to put them into channel.

This commit makes BoltProtocolV1 responsible for ChunkedOutput lifecycle. It
is created in constructor and closed when the protocol instance is closed.