Add Bookmark facility to Bolt server.
Bookmarks encode a transaction id.

We use them to support Causal Consistency, by supplying a bookmark
generated on one server to a subsquent transaction on a second
server in the same cluster. The second server will ensure that the
subsequent transaction sees a state at least as up to date as
the state seen on the first server, or throw an exception in case
of timeout.