Rebuild indexes that fail to open
If an index that should be online fails to open on startup we would not
be able to apply updates to it, and it would thus become inconsistent
with the database. Should the index then be possible to open after a
restart of the database we would end up with an inconsistent index.

The situation described above is now handled by treating the index as
permanently failed, and then taking actions to correct it. The first and
possibly most important action is to log the failure, something that was
not done before. Then we recognise that the only thing a user could do
to restore the index would be to either drop it and recreate it, or
stop the database, delete the index store files from disk, then restart
the database. By automating this second option we can avoid the need for
a restart. Since we are already in the startup phase when this happens,
we can simply treat the index as incompletely populated instead, forcing
the database to rebuild the index without any further intervention.