Ensure that IndexProxy returns the correct config
When performing recovery of a constraint index, the IndexProxy would
contain an IndexConfiguration object that said that it was a normal
index rather than a constraint index. This was because we did not want
the index to have to go through a tentative phase, but instead go to the
online state immediately.

This issue is solved by separating the configuration of the index from
whether it should flip through a tentative state or not.

There was nothing in the code that depended on this IndexConfiguration
object being correct after recovery, but it is possible that such code
could be written in the future...